% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_score.R
\name{adjust_score}
\alias{adjust_score}
\title{Title}
\usage{
adjust_score(score, outcome)
}
\arguments{
\item{score}{A scoreline}

\item{outcome}{A character indicating whether the match was won (W) or lost (L)}
}
\value{
a string
}
\description{
Title
}
\examples{
adjust_score("6-2", "6-2", "L")

}
