% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Eloratings_nolag.r
\name{Eloratings_nolag}
\alias{Eloratings_nolag}
\title{Creates elo ratings for players within a specified time frame without lag.}
\usage{
Eloratings_nolag(date1, date2, x)
}
\arguments{
\item{date1}{The beginning date for ratings.}

\item{date2}{The end date for ratings.}

\item{x}{A data frame where first column is the number of days from a specified date, the second column is Player 1, the third column is Player 2, and the last column is the result, either a 0 or 1.}
}
\value{
A list of elo ratings.
}
\description{
Creates elo ratings for players within a specified time frame without lag.
}

