% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Eloratings_status.r
\name{Eloratings_status}
\alias{Eloratings_status}
\title{Creates elo ratings for players within a specified time frame, also including matches one month before the specified time frame.}
\usage{
Eloratings_status(date1, date2, results)
}
\arguments{
\item{date1}{The beginning date for ratings, example 2015-03-14.}

\item{date2}{The end date for ratings, example 2015-03-14.}

\item{results}{A data frame where first column is the number of days from a specified date, the second column is Player 1, the third column is Player 2, and the last column is the result, either a 0 or 1.}
}
\value{
A list of elo ratings and match results.
}
\description{
Creates elo ratings for players within a specified time frame, also including matches one month before the specified time frame.
}

