% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Eloratingstodf.r
\name{Eloratingstodf}
\alias{Eloratingstodf}
\title{Transforms a list of elo ratings into a data frame.}
\usage{
Eloratingstodf(x)
}
\arguments{
\item{x}{A list of elo ratings.}
}
\value{
The elo ratings in a data frame.
}
\description{
Transforms a list of elo ratings into a data frame.
}

