% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournament_ids.R
\name{get_tournament_ids}
\alias{get_tournament_ids}
\title{Title}
\usage{
get_tournament_ids(year = 2022, tour = "atp")
}
\arguments{
\item{tour}{string indicating which tour; "atp" or "ch"}
}
\description{
Title
}
