% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatchResultselofunction.r
\name{MatchResultselofunction}
\alias{MatchResultselofunction}
\title{Formats an updated MatchStats database to be able to find elo ratings.}
\usage{
MatchResultselofunction(x)
}
\arguments{
\item{x}{Database of updated match statistics, namely the resulting database from the MatchStatsfunction.}
}
\value{
A dataset of updated match results formatted to find elo ratings.
}
\description{
Formats an updated MatchStats database to be able to find elo ratings.
}

