% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatchStatsfunction.r
\name{MatchStatsfunction}
\alias{MatchStatsfunction}
\title{Creates a database of match statistics for new tournaments and adds the statistics to the current database.}
\usage{
MatchStatsfunction(x)
}
\arguments{
\item{x}{A list of websites from the ATP World Tour website.  Use websites of the player draw from the Results Archive page.}
}
\value{
A complete database of match statistics, combining the built-in database with statistics from the new tournaments.
}
\description{
Creates a database of match statistics for new tournaments and adds the statistics to the current database.
}

