% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NewMatchStats.r
\name{NewMatchStats}
\alias{NewMatchStats}
\title{Creates a database of match statistics for tournaments.}
\usage{
NewMatchStats(website)
}
\arguments{
\item{website}{A list of websites from the ATP World Tour.  Use websites of the player draw from the Results Archive page.}
}
\value{
A complete database of match statistics.
}
\description{
Creates a database of match statistics for tournaments.
}

