% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Predictfunction_lag.r
\name{Predictfunction_lag}
\alias{Predictfunction_lag}
\title{Makes a prediction based on elo ratings, with lag, for a given time period.}
\usage{
Predictfunction_lag(player1, player2, elo)
}
\arguments{
\item{player1}{First player in the head to head matchup.}

\item{player2}{Second player in the head to head matchup.}

\item{elo}{A set of elo ratings, in the original elo ratings format.}
}
\value{
A probability that player 1 will win the tennis match.
}
\description{
Makes a prediction based on elo ratings, with lag, for a given time period.
}

