% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MatchResultselofunction.r
\name{MatchResultselofunction}
\alias{MatchResultselofunction}
\title{Formats a database of match statistics to be able to find elo ratings.}
\usage{
MatchResultselofunction(x)
}
\arguments{
\item{x}{Database of match statistics, in the form of the MatchStats database.}
}
\value{
A dataset of match results formatted to find elo ratings.
}
\description{
Formats a database of match statistics to be able to find elo ratings.
}

