% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Predict.r
\name{Predict}
\alias{Predict}
\title{Makes a prediction based on elo ratings for a given time period.}
\usage{
Predict(player1, player2, elo, results)
}
\arguments{
\item{player1}{First player in the head to head matchup.}

\item{player2}{Second player in the head to head matchup.}

\item{elo}{A set of elo ratings, in the original elo ratings format.}
}
\value{
A probability that player 1 will win the tennis match.
}
\description{
Makes a prediction based on elo ratings for a given time period.
}

